exporter/requirements.txt
nats-py
python-dotenv
dataclasses-json
aiohttp
psycopg2-binary
